apps/Server/requirements.txt
# =============================================================================
# Template Repository - Python Dependencies
# =============================================================================

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# HTTP Client
httpx>=0.25.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
email-validator>=2.0.0
